[STYLE] Create chatting ui
