+++ app/src/main/java/com/example/project_flow_android/ui/chat/fragment/ChatCreateFragment.kt
@@ -1,0 +1,19 @@
+package com.example.project_flow_android.ui.chat.fragment
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.project_flow_android.R
+
+class ChatCreateFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_chat_create, container, false)
+    }
+}
