+++ app/src/main/java/com/example/project_flow_android/ui/chat/fragment/ChatFragment.kt
@@ -1,4 +1,4 @@
-package com.example.project_flow_android.ui.chat
+package com.example.project_flow_android.ui.chat.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,13 +7,15 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.project_flow_android.R
 
-class UserInfoFragment : Fragment() {
+class ChatFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_user_info, container, false)
+        return inflater.inflate(R.layout.fragment_chat, container, false)
     }
+
+
 }
