+++ app/src/main/java/com/example/project_flow_android/ui/chat/fragment/ManageFragment.kt
@@ -1,4 +1,4 @@
-package com.example.project_flow_android.ui.chat
+package com.example.project_flow_android.ui.chat.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,15 +7,14 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.project_flow_android.R
 
-class ChatFragment : Fragment() {
+class ManageFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_chat, container, false)
+        return inflater.inflate(R.layout.fragment_manage, container, false)
     }
-
 
 }
