+++ app/src/main/res/layout/fragment_manage.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.chat.ManageFragment">
+    tools:context=".ui.chat.fragment.ManageFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
