+++ app/src/main/res/layout/fragment_user_info.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.chat.UserInfoFragment">
+    tools:context=".ui.chat.fragment.UserInfoFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
