+++ app/src/main/res/values/strings.xml
@@ -14,4 +14,8 @@
     <string name="manage_exit_room">방 나가기</string>
     <string name="user_info_attend_project">참여한 프로젝트</string>
     <string name="user_info_finish">완료</string>
+    <string name="new_chatting">새로운 채팅</string>
+    <string name="chat_create_project">프로젝트</string>
+    <string name="chat_create_user">참여원</string>
+    <string name="chat_create_start">시작하기</string>
 </resources>
